+++ python/tests/functional/eyes_selenium/test_usdk_commands.py
@@ -75,7 +75,7 @@ def test_usdk_commands_open_close_eyes(local_chrome_driver):
     assert test_result["appName"] == "USDK Test"
     assert test_result["name"] == "USDK Commands open close"
 
-    manager_close_result = commands.manager_get_results(mgr, False, 100)
+    manager_close_result = commands.manager_get_results(mgr, False, False, 100)
 
     assert manager_close_result == {
         "exceptions": 0,
@@ -113,7 +113,7 @@ def test_usdk_commands_open_abort_eyes(local_chrome_driver):
     assert test_result["appName"] == "USDK Test"
     assert test_result["name"] == "USDK Commands open abort"
 
-    manager_close_result = commands.manager_get_results(mgr, False, 100)
+    manager_close_result = commands.manager_get_results(mgr, False, False, 100)
 
     assert manager_close_result == {
         "exceptions": 1,
@@ -170,7 +170,7 @@ def test_usdk_commands_open_check_close_eyes(local_chrome_driver):
     assert test_result["appName"] == "USDK Test"
     assert test_result["name"] == "USDK Commands open check close"
 
-    manager_close_result = commands.manager_get_results(mgr, True, 100)
+    manager_close_result = commands.manager_get_results(mgr, True, False, 100)
 
     assert manager_close_result == {
         "exceptions": 0,