+++ python/tests/unit/eyes_selenium/test_runner.py
@@ -1,5 +1,5 @@
 import pytest
-from mock import call, patch
+from mock import ANY, MagicMock, call, patch
 
 from applitools.selenium import ClassicRunner, RunnerOptions, VisualGridRunner
 from applitools.selenium.__version__ import __version__
@@ -37,6 +37,20 @@ def test_visual_grid_runner_creation_test_concurrency():
         assert get_instance.mock_calls == [
             call(SeleniumWebDriver, "eyes.selenium.visualgrid.python", __version__),
             call().core_make_manager(ManagerType.UFG, concurrency=3),
+        ]
+
+
+def test_visual_grid_runner_creation_remove_duplicate_tests():
+    get_instance = "applitools.selenium.command_executor.CommandExecutor.get_instance"
+    command_executor_mock = MagicMock()
+    with patch(get_instance, return_value=command_executor_mock) as get_instance:
+        command_executor_mock.manager_get_results.return_value = {}
+        runner = VisualGridRunner(RunnerOptions().remove_duplicate_tests())
+        runner.get_all_test_results(False)
+
+        assert command_executor_mock.mock_calls == [
+            call.core_make_manager(ANY, concurrency=5),
+            call.manager_get_results(ANY, False, True, 1800),
         ]