+++ python/eyes_images/CHANGELOG.md
@@ -1,4 +1,4 @@
-## [vNext]
+## [5.18.0] - 2023-04-26
 ### Add
 - DOM can now be attached to images.Eyes checks via `dom` method of target [Trello 3196](https://trello.com/c/FZ6KcVd6)
 - Runners now support `remove_duplicate_tests` option via RunnerOptions and `set_remove_duplicate_tests` method [Trello 3202](https://trello.com/c/tfg6cLPd)